flask/requirements.txt
amqp==2.1.4
aniso8601==1.3.0
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-Login==0.4.0
Flask-PyMongo==0.5.1
Flask-RESTful==0.3.6
funcsigs==1.0.2
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
jsonrpcserver==3.5.2
jsonschema==2.6.0
MarkupSafe==0.23
pika==0.10.0
pymongo==3.5.1
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
redis==2.10.6
requests==2.18.3
six==1.11.0
urllib3==1.22
vine==1.1.4
Werkzeug==0.12.2
